+++ src/Rector/SetList.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\CodeStyle\Rector;
+
+final class SetList
+{
+    public const YII_CORE = __DIR__ . '/sets/yii-core.php';
+}