+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## 1.0.1 under development
 
-- Enh #4: Add the `no_empty_statement` rule to `@Yiisoft/Core` (@mspirkov)
+- New #4: Add the `no_empty_statement` rule to `@Yiisoft/Core` (@mspirkov)
+- New #11: Add `RemoveOverrideAttributeRector` Rector rule (@vjik)
+- New #11: Add `SetList::YII_CORE` Rector set (@vjik)
 
 ## 1.0.0 January 23, 2026